feat(app): split danger zone into dedicated settings tab
Group org + account destructive actions into /settings/danger
(Railway-style). Each Card uses a destructive Alert callout with
AlertTriangleIcon and a self-aligned destructive button. Tab gets
the AlertTriangle icon. Route lives under settingsLayout (not
_org-scope) so account-delete stays reachable without an active org.
