+++ apps/app/src/features/danger/danger.route.tsx
@@ -1,0 +1,8 @@
+import { createRoute, lazyRouteComponent } from "@tanstack/react-router";
+import { settingsLayout } from "../../router/layouts";
+
+export const dangerRoute = createRoute({
+  getParentRoute: () => settingsLayout,
+  path: "danger",
+  component: lazyRouteComponent(() => import("./danger.page"), "DangerPage"),
+});
