+++ apps/app/src/shared/components/contextual-tabs.tsx
@@ -2,11 +2,13 @@ import type { OrgPermissions } from "@packages/access-control";
 import { NavLink } from "@packages/ui/components/ui/nav-link";
 import { cn } from "@packages/ui/libs/utils.js";
 import { Link, useRouterState } from "@tanstack/react-router";
+import { AlertTriangleIcon, type LucideIcon } from "lucide-react";
 import { useAuthorization } from "../auth/use-authorization";
 
 interface TabItem {
   to: string;
   label: string;
+  icon?: LucideIcon;
   requires?: OrgPermissions;
   requiresOrg?: boolean;
 }
@@ -26,6 +28,7 @@ const SETTINGS_TABS: readonly TabItem[] = [
     requiresOrg: true,
   },
   { to: "/settings/account", label: "Account" },
+  { to: "/settings/danger", label: "Danger", icon: AlertTriangleIcon },
 ];
 
 interface ContextualTabsProps {
@@ -47,11 +50,17 @@ export function ContextualTabs({ className }: ContextualTabsProps) {
       aria-label="Settings sections"
       className={cn("flex items-center gap-1 overflow-x-auto", className)}
     >
-      {visibleTabs.map((tab) => (
-        <NavLink key={tab.to} variant="underline" active={pathname === tab.to} asChild>
-          <Link to={tab.to}>{tab.label}</Link>
-        </NavLink>
-      ))}
+      {visibleTabs.map((tab) => {
+        const Icon = tab.icon;
+        return (
+          <NavLink key={tab.to} variant="underline" active={pathname === tab.to} asChild>
+            <Link to={tab.to} className="gap-1.5">
+              {Icon && <Icon className="size-3.5" />}
+              {tab.label}
+            </Link>
+          </NavLink>
+        );
+      })}
     </nav>
   );
 }
